made starter preview call api for content (yet it is still stub); cleanups
