+++ ui/src/components/ConfigurationForm/__tests__/ConfigurationForm.component.test.tsx
@@ -1,6 +1,7 @@
-import { render, screen, within, waitFor } from '@testing-library/react';
+import {render, screen, waitFor, within} from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import { ConfigurationForm } from '../ConfigurationForm.component';
+import {BrowserRouter} from 'react-router-dom';
+import {ConfigurationForm} from '../ConfigurationForm.component';
 
 global.fetch = jest.fn();
 
@@ -13,7 +14,7 @@ describe('ConfigurationForm component', () => {
     );
 
     const user = userEvent.setup();
-    render(<ConfigurationForm />);
+    render(<BrowserRouter><ConfigurationForm /></BrowserRouter>);
 
     // when
     await user.type(screen.getByRole('textbox', { name: /Project name/i }), 'test-project');
@@ -74,7 +75,7 @@ describe('ConfigurationForm component', () => {
   test('for validation error and lack of network request while trying to submit not complete form', async () => {
     // given
     const user = userEvent.setup();
-    render(<ConfigurationForm />);
+    render(<BrowserRouter><ConfigurationForm /></BrowserRouter>);
 
     // when
     await user.type(screen.getByRole('textbox', { name: /Project name/i }), 'test-project');
@@ -108,7 +109,7 @@ describe('ConfigurationForm component', () => {
   test('reset flow to form default values', async () => {
     // given
     const user = userEvent.setup();
-    render(<ConfigurationForm />);
+    render(<BrowserRouter><ConfigurationForm /></BrowserRouter>);
 
     // when
     await user.type(screen.getByRole('textbox', { name: /Project name/i }), 'test-project');
@@ -153,7 +154,7 @@ describe('ConfigurationForm component', () => {
   test('effect implementation field reset flow based on scala version change', async () => {
     // given
     const user = userEvent.setup();
-    render(<ConfigurationForm />);
+    render(<BrowserRouter><ConfigurationForm /></BrowserRouter>);
 
     // when
     await user.click(screen.getByRole('button', { name: /Effect type/i }));
@@ -183,7 +184,7 @@ describe('ConfigurationForm component', () => {
   test('effect implementation field reset flow based on effect type change', async () => {
     // given
     const user = userEvent.setup();
-    render(<ConfigurationForm />);
+    render(<BrowserRouter><ConfigurationForm /></BrowserRouter>);
 
     // when
     await user.click(screen.getByRole('button', { name: /Effect type/i }));
@@ -212,7 +213,7 @@ describe('ConfigurationForm component', () => {
   test('json field reset flow based on effect type change', async () => {
     // given
     const user = userEvent.setup();
-    render(<ConfigurationForm />);
+    render(<BrowserRouter><ConfigurationForm /></BrowserRouter>);
 
     // when
     await user.click(screen.getByRole('button', { name: /Effect type/i }));
